Fix long-to-int cast in LiteralInterpreter
ConstantExpression takes long as its value for some numeric types. It
should avoid direct cast to integers. The bug was introduced with the
LiteralInterpreter but got exposed during the "Assignment" refactoring.

+++ presto-main/src/main/java/com/facebook/presto/sql/planner/LiteralInterpreter.java
@@ -156,7 +156,7 @@ public final class LiteralInterpreter
             return new SqlIntervalDayTime((long) node.getValue());
         }
         if (type instanceof IntervalYearMonthType) {
-            return new SqlIntervalYearMonth((int) node.getValue());
+            return new SqlIntervalYearMonth(((Long) node.getValue()).intValue());
         }
 
         // We should not fail at the moment; just return the raw value (block, regex, etc) to the user
